attempts to make the buttons work properly
the prev/next and content are put into one flexbox-container in order to prevent overlaps. this works, although the prev/next buttons no longer scroll with the content (how do you get flexbox divs to sit at the bottom of the viewport?)
works on notepage, does not work as effectively on poempage. feel free to revert if this doesn't help.